refactoring - view render
