--- src/ts/store/action.ts
@@ -1,11 +1,0 @@
-enum Actions {
-  init = 'init',
-  reset = 'reset',
-  setCarNames = 'setCarNames',
-  setTotalAttempts = 'setTotalAttempts',
-  race = 'race',
-  raceFinished = 'raceFinished',
-  notifyWinner = 'notifyWinner',
-}
-
-export default Actions
