+++ src/ts/view/playboard.ts
@@ -1,6 +1,6 @@
+import { State } from '../types.js'
 import el from '../util/dom.js'
 import View from './constructor.js'
-import { State } from '../store/index.js'
 import Player from './player/index.js'
 
 type WatchState = Pick<State, 'cars'>
@@ -19,8 +19,7 @@ export default class Playboard extends View {
       const $entries = cars.length
         ? (cars.map((name, i) => el(`<racingcar-player name=${name} index=${i}>`)) as Player[])
         : []
-      el(this, $entries)
-      this.show()
+      this.render($entries).show()
     }
   }
 }
