+++ src/ts/view/winner.ts
@@ -1,19 +1,18 @@
+import { State, Actions } from '../types.js'
 import el from '../util/dom.js'
 import View from './constructor.js'
-import Actions from '../store/action.js'
-import { State } from '../store/index.js'
 
 type WatchState = Pick<State, 'winners'>
 export default class Winner extends View {
   static #template = /* html */ `
-  <section class="d-flex justify-center mt-5">
-    <div>
-      <h2>🏆 최종 우승자:  🏆</h2>
-      <div class="d-flex justify-center">
-        <button type="button" class="btn btn-cyan">다시 시작하기</button>
+    <section class="d-flex justify-center mt-5">
+      <div>
+        <h2>🏆 최종 우승자:  🏆</h2>
+        <div class="d-flex justify-center">
+          <button type="button" class="btn btn-cyan">다시 시작하기</button>
+        </div>
       </div>
-    </div>
-  </section>
+    </section>
   `
 
   $winner
@@ -25,7 +24,7 @@ export default class Winner extends View {
     this.$winner = $container.querySelector('h2') as HTMLHeadingElement
     this.$restart = $container.querySelector('button') as HTMLButtonElement
     this.$restart.addEventListener('click', this.onRestart)
-    el(this, [$container])
+    this.render($container)
   }
 
   watch = ({ winners }: State): WatchState => {
